[REFACTOR][mumi]: StoreService, StoreServiceImpl ->  StoreCommandService, StoreQueryService로 통합
